+++ xbmc/addons/kodi-dev-kit/include/kodi/c-api/addon-instance/imagedecoder.h
