+++ xbmc/addons/kodi-dev-kit/include/kodi/versions.h
@@ -105,7 +105,8 @@
 #define ADDON_INSTANCE_VERSION_IMAGEDECODER           "2.1.1"
 #define ADDON_INSTANCE_VERSION_IMAGEDECODER_MIN       "2.1.0"
 #define ADDON_INSTANCE_VERSION_IMAGEDECODER_XML_ID    "kodi.binary.instance.imagedecoder"
-#define ADDON_INSTANCE_VERSION_IMAGEDECODER_DEPENDS   "addon-instance/ImageDecoder.h"
+#define ADDON_INSTANCE_VERSION_IMAGEDECODER_DEPENDS   "c-api/addon-instance/imagedecoder.h" \
+                                                      "addon-instance/ImageDecoder.h"
 
 #define ADDON_INSTANCE_VERSION_INPUTSTREAM            "3.0.2"
 #define ADDON_INSTANCE_VERSION_INPUTSTREAM_MIN        "3.0.1"
